requirements.txt
streamlit
opencv-python-headless
numpy
pytesseract
Pillow
reportlab